refactors userblocks into HHB locations
